Reduce Interactive Compiler Dependencies
Reorganized the code a bit to reduce the dependency on the interactive compiler. This makes it easier to reuse parts of the library with a non-interactive compiler.

Users that don't have an interactive compiler have to implement the common.CompilerAccess trait and mix it into their Refactoring instances.

For existing users, nothing should change as MultiStageRefactoring extends from InteractiveScalaCompiler, which itself extends CompilerAccess.