Organize Imports that adds Imports
The first implementation for an organize imports that finally does what
JDT developers expect: missing imports are added!

Because this requires more information about the code than the refactoring
library has, it is implemented as follows:
  - Preparing the refactoring returns a list of strings with all the missing imports names.
  - When performing an organize imports, there's a new parameter that holds tuples of ("package.declaration", "TypeName") that will be added as new import statements.

In short, the IDE has to do all the hard work figuring out which imports to add, e.g. by asking the user.

Default values are in place so that this change should not break any existing code.