Make scala-refactoring compatible with unit-unloading of askLoadedTyped
The new behavior of unit unloading is introduced in https://github.com/scala/scala/pull/3158
on which this commit is strictly dependent.

In the longer term, scala-refactoring depends on a global "load-all-before-index" and "remove-all-after" semantics.

It would be wise  to refactor scala-refactoring to load files parcimoniously throughout the index building process. But this is already a step in the right direction.