Print type keyword when the type of an object should be printed
I would prefer to fix this somewhere in the printing logic and not in
the construction logic of the tree but I didn't find a way to
differentiate the Ident trees correctly in the printer. For

  object X { def x = 0 }

the Ident trees of `X.type` and `X.x` seem to have the same shape at the
beginning.

Fixes #1002233