Add return type to ValOrDefDef with single expressions in braces
Fixing this was complicated because the compiler removes braces from the
tree when the code would compile without them. Therefore we manually
have to restore them.

To get the braces (and whitespace) back, we search for an equal sign
in the area of the DefDef (which indicates the start of the rhs). Once
it is found, all contents between the equal sign and the start and the
expression of the rhs are put back into the tree.

Because an equal sign can also occur in a comment, we have to parse the
region instead of simply looking for the equal sign. This adds some
overhead, but hopefully not too much.

Furthermore, as reviewers suggested, some variable names are renamed.

This also fixes a bug in the test suite where the source file is passed
to the refactoring logic.