Revert "Remove Eclipse update-site and feature"
This reverts commit 9470fa61e1a06a1f9f29588175f2b3163e760d52.

As it turned out, I was wrong. The IDE very well depends on the
update-site. Its maven build needs the update site to find the
scala-refactoring dependency. The better fix is to not stress
scala-refactoring with the update-site but to fix the Scala IDE build.
Until this is done, the change is reverted in order to not block the
current IDE builds.