Reintroduce org.scala-refactoring.library project
I don't know any better than to do this. The project itself has no
content beside from the maven pom. This seems to be necessary because
maven doesn't allow to build the contents in the parent pom. Therefore I
added some symlinks to the files and directories in the root directory.

Some plugin definitions from the root pom.xml are also moved to the
sub module pom.xml because otherwise the top level src-directory would
have been included in the compilation process of the main module (and
not in the compilation process of the sub module is it is intended).

These changes are only temporarily. Once the Scala IDE build is improved
in so far that it no longer needs the scala-refactoring update site, we
can revert this commit.