Don't duplicate imports while grouping them
This commit reimplements the import grouping logic in a self contained, easily
testable function, and thereby fixes a bug, that led to the duplication of
imports if one import group was the prefix of another import group.

Note that the unit test that is added by this commit lives in exactly the same
package as the code it covers, since this allows us to access package private
APIs. We should consider doing the same for the rest of the tests, that
currently live under `scala.tools.refactoring.tests`.

Fixes #1002526