Do not add opening paren in case the document has errors
This case is similar to the case where a closing paren is missing. The
only internal difference is that the compiler actually thinks that the
document is a valid tree. The compiler creates a method symbol out of
the broken method declaration, assigns a Null type as the return value
and skips the rest of the line completely. The tree does not show that
there is a syntax error in the sources. By checking if the return type
is Null we should be on the safe side in assuming that there is an error
because users generally don't use that type anywhere.