Avoid tree printing in organize imports logic
Tree printing is buggy and in this case it led to an exception because
some synthetic tree of a for comprehension was printed. In order to
avoid the exception we no longer call the tree printing logic at all.
Instead we compare symbols and their names, which leads along the way
to even better performance.

Fix #1002476