Implement backtracking support for `SourceWithMarker`
Up until now, the movements related to `SourceWithMarker` did not implement
backtracking. As a consequence, `'a'.zeroOrMore ~ 'a'` would not consume "a",
because `'a'.zeroOrMore` would always move over "a", leaving no room for the
following movement. While this wasn't a severe restriction for moving forward
since rules can be adapted accordingly, this becomes a real problem when moving
back, since `('a' ~ 'a'.zeroOrMore).backward` applies `'a'.zeroOrMore` before
`'a'`.