Add empty lines around added imports and put import at right location
This commit is doing two things. For every import that is added, it is
checked that it is surrounded by empty lines if they are necessary.
"Necessary" means that the empty lines are added in two cases: After a
package declaration and before a class declaration. Once an import is
added, a code snippet like the following

    package pkg
    trait X {
      def d: Try[Int]
    }

will look like this:

    package pkg

    import scala.util.Try

    trait X {
      def d: Try[Int]
    }

The added empty lines actually complicate the implementation and are not
needed to fix the problem of the missing import. But they are needed to
ensure consistency in the implementation. If we have a code snippet like
the following

    package pkg

    trait X {
      def d: Try[Int]
    }

The question about where to put the import declaration comes up. Shall
it be added before or after the empty line? Or shall the empty line even
be replaced by the import declaration? To me, it looked easier to just
forget about these questions and instead to ensure that empty lines
always exist around the import.

In case some imports already exist, there is no need for an import
declaration:

    package pkg

    import scala.util.Try

    trait X {
      def d: Try[Int]
      def s: Success[Int]
    }

will turn into

    package pkg

    import scala.util.Try
    import scala.util.Success

    trait X {
      def d: Try[Int]
      def s: Success[Int]
    }

without an additional line added before the new import declaration.

The second problem that is fixed by this commit is that the import is
now added before a comment:

    package pkg
    /**
     * Scaladoc
     */
    trait X {
      def d: Try[Int]
    }

will turn into

    package pkg

    import scala.util.Try

    /**
     * Scaladoc
     */
    trait X {
      def d: Try[Int]
    }

Previously it was added after the comment, which is not what anyone
would expect.

Fixes #1002672