Make `Movement.||` greedy
Up until now, `Movement.||` used a short circuiting logic, like a normal `or`
operator. This meant for example, that `'ab' || 'abb'`, applied to "abb",
would not have consumed the entire string, but only "ab". Applying the same
movement, but backward, starting at the end, however consumed everything,
since 'ab' cannot be applied to the end of "abb". Likewise, it is easy to
construct similar examples, where going forward consumes only a part, and
going back the whole string.

Most importantly, this asymmetry affected `Movement.id`, and therefore the
correctness of refactoring operations relying on it. Thus I decided
to enforce symmetry by making `||` go for the longest instead of the first
match.